feat: remove the need to add `.svelte` to the end of indiviual component imports
Alongside every component their is a `.js` and `.d.ts` file with the same name. The Javascript file simply exports the svelte component. Like this:

    import AB from "./AB.svelte";
    export default AB;

The accomponying Typescript file looks like this:

    import { SvelteComponentTyped } from "svelte";
    export default class AB extends SvelteComponentTyped<{
        color?: string;
        size?: string | number;
        strokeWidth?: string | number;
    }> {}

---

This allowes for better type checking for when only one component is imported.

Another small bonus is that instead of typing:

    import CurrencyBitcoin from "tabler-icons-svelte/icons/CurrencyBitcoin.svelte"

It can be shortened to:

    import CurrencyBitcoin from "tabler-icons-svelte/icons/CurrencyBitcoin"

---

I don't know of the is going to be worth it because it does add some weight to the package.